write_uploaded_file: add optional overwrite parameter.
If overwrite is true, attempt to replace an existing file (similar to
the old behavior of the function.)  If overwrite is false, don't do
that (and if the file already exists, a FileExistsError will be
raised.)

(Note that currently, UploadFilesForm tries to avoid calling
write_uploaded_file if the file already exists.)

+++ physionet-django/project/utility.py
@@ -176,15 +176,16 @@ def readable_size(num, suffix='B'):
     return '{:.1f}{:s}{:s}'.format(num, 'Y', suffix)
 
 
-def write_uploaded_file(file, write_file_path):
+def write_uploaded_file(file, write_file_path, overwrite=True):
     """
     file: request.FILE
     write_file_path: full file path to be written
     """
-    try:
-        os.unlink(write_file_path)
-    except FileNotFoundError:
-        pass
+    if overwrite:
+        try:
+            os.unlink(write_file_path)
+        except FileNotFoundError:
+            pass
     with open(write_file_path, 'xb') as destination:
         for chunk in file.chunks():
             destination.write(chunk)
